multi_challenge/requirements.txt
pydantic==2.10.6
python-dotenv==1.0.1
torch==2.5.1
tqdm==4.66.2
transformers==4.44.1
openai==1.53.0